FAB-7382 Remove ReadyChan() from  Deliver ledgers
This CR reworks the ledger interfaces and implementations
used by the Deliver service to remove the ReadyChan() function,
which has not been used since FAB-7273 was merged. The
chainSupport struct for the peer is also updated to only
include the ledger. The file ledger is now returned when
Reader() is called by the deliver service.

FAB-7382 #done

Change-Id: I6d08f30b9093ec3120cdeee749aa51afe54e2cdc
Signed-off-by: Will Lahti <wtlahti@us.ibm.com>