[FAB-11701] Prover Peer proto (with import request)
Define the protobuf messages and services needed to allow a
client to request the execution of specific command,
(i.e.: request import, transfer, list tokens, and so on).
 The messages should guarantee the following properties:
- client request and prover peer response are authenticated
- prover peer response is cryptographically bound to the
corresponding request. The link must be succinct and it
should not be possible to derive from it the response's content.

As example of the available commands, an import request
command will be defined as well.

Change-Id: I4d4bade5e73aa1b6c89a9dbb0103195eed479a03
Signed-off-by: Angelo De Caro <adc@zurich.ibm.com>