FAB-11506 configtxlator do not deeply encode nil
The protolator code used by configtxlator checks to see if the proto
message contains opaque fields (ie, byte slices which are really proto
messages).  However, sometimes, a field is deliberately omitted, and
there is not enough information to determine the encoding of the field,
unless it is populated.  This can result in errors encoding sparse
messages as protobuf code is told to emit defaults.

This CR adds a check before attempting to deeply marshal a field (be it
plain, map, or slice), that if the field is nil, it should not be
deeplly marshaled.

This addresses a bug seen decoding channel configs which embed the
sparse config update structure.

Change-Id: I72c76c92dfb4ad3e5275b2da16507de29c1d4822
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>