[FAB-11569] Fix public/sideDB updates final commit
Currently, when submitting the commits of both public updates and sideDB
updates into the database, they are done in the same
`s.VersionedDB.ApplyUpdates(updates.PubUpdates.UpdateBatch, height)` call.

However, the variable in use is PubUpdates, which actually includes both pub
and private data.

Hence after discussion with manish, a quick fix to this is to introduce a
combined updates result.

Change-Id: Idebd92ada1080d91d6aab9da2d1da002acda30d1
Signed-off-by: Baohua Yang <yangbaohua@gmail.com>