[FAB-11574] Use AdminClient to create topics
The Kafka AdminClient API is now used in order
to create the topic required for a channel.
Prior to this, the orderer relied on the Kafka
cluster to have auto create topics enabled.

This API is only supported for Kafka v0.10.1.0
and higher.

If the API fails, the code still falls through
to the auto create path.

Change-Id: Id23d5b5f5826b24cb0cc3202e3ac6088e5ae3a49
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>