FAB-11587 Add new tar gz chaincode package format
Historically, chaincodes have been package inside of a protobuf, with a
nested tar or other archive inside.  This makes inspecting packages by
users very challenging.  It also excludes packaging code through any
mechanism other than the fabric tools very challenging.

This adds a new chaincode package format which expect a .tar.gz format
set of bytes, which contains exactly two files:

 - Chaincode-Package-Metadata.json - Specifies the type and path
 - <any other file> - Contains the code package archive, usualy tar

In the future, it would be nice to support more of a buildpack style
packaging, where the tar does not necessarily require a manifest, but
instead infers the operational parameters based on the package layout.
However, as an intermediate step, this preserves the majority of the
existing function, while moving us in the right direction.

Makes #FAB-11587 Done

Change-Id: I06e049dd38164f91f194ace80fe8e7c5b10b5f09
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>