FAB-11588 Have package-provider parse fs bytes
The package provider currently returns only the bytes stored on the
filesystem.  Since we are encoding these bytes as a tar with a manifest
file, we need to parse them and only return the embedded code package.

This also obviates the need for a ChaincodeInstallPackage proto, so this
CR removes it.

Makes #FAB-11588 Done

Change-Id: I18226576cb221afa5cc664492a8e1e3331066ad2
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>