FAB-13500 Add serializer for structs to db
The v1.x implementation of LSCC stored all of the chaincode information
(except for collections) in a single monolithic value using protobuf.
This has the problem that any change to the chaincode definition
invalidates all in-flight requests, so for instance, changing the
endorsement policy (which has no affect on the output of the tx) will
invalidate a transaction unnecessarily.

Therefore, it's desirable to have a way to store individual structure
fields as keys, but without introducing the complexity of serializing
all fields individually.

Change-Id: I46a30f34f1e7fa48e84e5029b0d19c0409348858
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>