[FAB-11646] Add interceptors to comm.GRPCServer
Adds the ability to set unary and stream
interceptors when configuring comm.GRPCServer
instances.

Adds the grpc-middleware library in order to
allow chaining of interceptors.
NOTE: had to remove unused package folders
manually and add grpc-middleware to the
noverify list due to hashes not matching for
symlinks across OS versions.

Change-Id: I0c9baa570ea84aa4be8377dedd0ada30c7e7a291
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>