[FAB-11657] Fix Testviewer URL in docs Readme
The location url of the Testviewer application has
changed. To limit future churn here in fabric
should the location change again, we modify the
link in this doc to direct the reader to the
Testviewer Readme file in the fabric-test repo
where we will maintain the only actual direct link.

Also, we add and update links to the tests results
summary pages in jenkins.

FYI: Refer also to accompanying URL change in the
Testviewer directory in fabric-test repo:
https://gerrit.hyperledger.org/r/#/c/25807/

Change-Id: I31e84e65d6974d7fad6246f00ca9775d0aebf8a0
Signed-off-by: Scott Zwierzynski <scottz@us.ibm.com>