[FAB-11719] Better Markup of Document
FAB-11719 #done

The first paragraph of "Building Your First Network" of section of the
document is markuped as follows:

    The ``configtxgen tool`` is ...

It is better to be markuped as follows since only a name of
file/command is quoted in other part of the document page.

    The ``configtxgen`` tool is ...

This patch updates the markup as above.

Change-Id: I73410da896cfea6e4743a72e0b78a623edf869b8
Signed-off-by: Yuta Namiki <y-namiki@ak.jp.nec.com>