recon: commit only validTx's pvtData in pvtStore
Currently, gossip constructs pvtData only for validTx
as per VSCC. Hence, we are not storing all pvtData of
a block.

In the ledger, currently, we are storing pvtData of
invalidTx (i.e., failed MVCC checks) too. To be
consistent, this CR can stores the pvtData of only
validTx (as per MVCC check).

FAB-11805 #done

Change-Id: Ic4b228cade0456bc1ddc4ab80595e9e82d7aaa25
Signed-off-by: senthil <cendhu@gmail.com>