[FAB-16302] Update historydb key format
The historydb code assumes that the keys do not contain
a nil byte. In the presence of such keys, a historydb query
may include wrong results due to key collision.

This CR fixes this issue. However, it makes a slight change
in the format of the keys that are persisted and hence will
require dropping and rebuilding of existing historydb.

Change-Id: I8b451fff10e15e22948b01f18fddb16911866d76
Signed-off-by: manish <manish.sethi@gmail.com>
Signed-off-by: Wenjian Qiao <wenjianq@gmail.com>