[FAB-11902] fix deadlock in key-level validation
Key-level validation establishes a producer/consumer pattern, where
 * the producer is the goroutine performing validation of a transaction
   that modifies the key-level validation params of a key, signalling its
   validation result
 * the consumer is the goroutine performing validation of a transaction
   that retrieves the key-level validation params for the same key
A deadlock existed in the code because the consumer was holding a lock
(namely, validationContext.mutex, acquired in waitForValidationResults)
while entering the sleep state to wait for the producer (in
waitForAndRetrieveValidationResult). The producer on the other hand was
prevented from signalling the validation result and broadcast the
conditional variable that would have caused the consumer's sleep state to
be interrupted, because it needed to obtain the same
validationContext.mutex.

The problem has been solved by changing waitForValidationResults in the
following way: validationContext.mutex is held for the time required to
safely read from the map and build a local slice of dependency objects. The
mutex is then released and consumption of the validation results can
proceed with the local slice.

Furthermore, an investigation of the code for other deadlocks didn't reveal
any but lead to the removal of an unused variable.

Change-Id: Ibf2e53959416e29c770196a1a2f8f0dc05363bd7
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>
Signed-off-by: Matthias Neugschwandtner <eug@zurich.ibm.com>