code hygiene: rm marble examples from ledger
A long time ago in a galaxy far far away, we added marble example
app for ledger integration test (due to lack of availability of
other components in the beginning of v1.0 development).
Given that now we have sophisticated ledger level integration test
and all components integration test, marble example looks redundant and
hence can be removed from the ledger folder. Anyway, we have not
added any new feature to this part of code in the last 2 years.

FAB-12035 #done

Change-Id: I547d4f9d473bf8b2b6caef9d32f32bd9f40f1cba
Signed-off-by: senthil <cendhu@gmail.com>