Add function to get pvtdata metadata by keyhash
During validation, key-based endorsement needs to
check the endorsement based on the key hash.

In the current code, query executor exposes a function
to retrieve the private data metadata by the private raw key.
This function is meant to be used by the chaincode in a simulation.

At the validation time, this raw private key is not available to all
the peers and in the hashed rwset, only the hash of the key is available.
Hence query executer needs to expose a function that accepts
the hash of the key.

FAB-12086 #done

Change-Id: I7398865f6f28a919daff8095291492832976f4ff
Signed-off-by: manish <manish.sethi@gmail.com>