[FAB-12209] Speedup TestShuffle in discovery
The test TestShuffle uses a slice of 1 million elements in order the
shuffling to fail with a negligible probability.

This runs in 2 seconds without race detector, but -
with race detector enabled it runs very very slow (20 seconds).

I changed the million to a thousand, and the number of
permutations should be still high enough to have the test
fail with low probability.

FAB-12209 #done

Change-Id: I57ccfbe0051055bb470f1c87d44eb813ecd6196e
Signed-off-by: yacovm <yacovm@il.ibm.com>