[FAB-12287] handle successful decode of bad data
LSCC handles CDSPackage and SignedCDSPackage instances in pretty much
the same way. The current pattern seems to be to try to decode data from
lifecycle into CDSPackage and, if it works, use it; if it doesn't try to
decode it into a SignedCDSPackage.

That would be great if protobuf were guaranteed to fail decoding the
wrong object but it doesn't.

In preparation for an upgrade to the proto package, the CDSPackage
validation needs to be extended to catch more scenarios where bad data
has been "successfully" unmarshaled into an object. This needs to be
done in a way that increases confidence in the decoding process without
duplicating validation that lives in lifecycle.

Change-Id: Ib53ddf1260a08e2f2178ce0f8862da5429aff87b
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>