[FAB-12443] Tolerate CouchDB 2.2 Changes
CouchDB 2.2 does not throw an exception anymore
if a specified index does not exist.
This is a change and unit tests will need to not check
for exceptions if a named index in the query does not exist.

Example:
{"selector":\{"owner":"tom"}
, "use_index":["indexOwnerDoc","indexOwner"]}

If the document indexOwnerDoc or the index indexOwner does not exist,
an exception will not be thrown.

CouchDB unit tests are changed to reflect the changes.

Change-Id: I627a500599e696d580d06fe92fe8362f6923440e
Signed-off-by: Chris Elder <chris.elder@us.ibm.com>
(cherry picked from commit c8401e4c34cb54b5e91010021ce588e00d774f56)