[FAB-12614]MockPeerChaincode receives channel name
Chaincode's MockStub has an API "InvokeChaincode", which is used
for unit tests invoking other chaincode. This function converts
chaincodeName to a composite name internally:
chaincodeName = chaincodeName + "/" + channel.
However, this conversion is not applied to "MockPeerChaincode",
which registers another Mockstub. When writing unit tests
invoking chaincode on other channel, developers have to follow
the naming rule manually. In addition, the rule is not described
in the API usage.

This CR updates "MockPeerChaincode" to receive channel name as
a parameter for CC2CC on another channel.

FAB-12614 #done

Change-Id: Ifcabc19bd78cbaa8bcc702c403da112096b16643
Signed-off-by: Yuki Kondo <yuki.kondo@hal.hitachi.com>