[FAB-12534] Move Step() to its own logger
The Step() RPC is invoked by etcdraft's module on a very very
frequent basis (a few times per second).

This means that if the cluster logging module is on debug level,
it's very hard to read the logs and they grow up very fast.

As a countermeasure, I'm giving the Step() its own logging package.

Change-Id: Ib03f3bee7307879e6a30675014bc4317cb7883f4
Signed-off-by: yacovm <yacovm@il.ibm.com>