[FAB-12550] OSN config block hash chain validation
This change set adds support for hash chain validation in presense
of configuration blocks that are found inside the block batches
being validated.

This is important because if a config block alters the validation rules
inside a batch, then validating the signature/certificates of the last
block in the batch needs to take into account the config block updates
themselves.

Change-Id: I03ef2a268ddc9ff63755d3a878f77d4814c5279d
Signed-off-by: yacovm <yacovm@il.ibm.com>