[FAB-12591] Block puller out of a config block
This change set adds a function that creates a block puller
out of a configuration block.

It is supposed to be used to prime the block puller for the system
channel, in order to pull all blocks from the system channel.

The TLS CAs and endpoints are going to be extracted
from the config block.

Change-Id: I2c290e703ab9167dc8cd0643e2d3ff1f6bed35e3
Signed-off-by: yacovm <yacovm@il.ibm.com>