Change signature of internal evaluator function
Towards a refactoring of the validation/committer, we need to decouple
signature/identity validation and policy checking. This change-set
enables this by changing the signature of the function that internally
evaluates the policy: instead of directly invoking functions to
deserialize the supplied identity, checking compliance of the identity
with an MSP principal and verifying a signature, it calls functions
of a suitable interface. Implementers of the interface are then free
to perform any appropriate precomputation.

FAB-12774 #done

Change-Id: I9df76f801a6b26f303070899ee50f0b699f06e36
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>