recon: convert missingData slice to a map
For any internal usage, it is good to have map so that lookup and
comparison code can be written with lesser lines of codes. If the
order of the stored entries do not matter, then slice is not needed
and adds a lot of complexity during lookups.

Given above, this CR changes the slice used for missingData to a map.

In the BlockAndPvtData, we have Block, BlockPvtData, and Missing.
These names are not very informative and get confused with another
struct named BlockPvtData. Hence, we rename
BlockAndPvtData.BlockPvtData to BlockAndPvtData.PvtData

FAB-12895 #done

Change-Id: I60af0520fc028392703a38e46c5c15ac2707a139
Signed-off-by: senthil <cendhu@gmail.com>