[FAB-12882] Fix UT flake TestBlockPullerClone
In tests where we have 2 block pullers, or that the block
puller initiates a disconnect, the server side deliver service
may send 2 blocks down the stream before it senses the disconnect.
This interferes with the assertions of the next block pull after the
disconnect.

This change set fixes this by making the server side expect
a disconnect after the last block of that client is expected
to be pulled.

Change-Id: Ic5501098ced12684f95d4a4f698f06963d6f7908
Signed-off-by: yacovm <yacovm@il.ibm.com>