[FAB-12919] Serialize `SendSubmit` invocation.
This CR adds a mutex to guard parallel access to `SendSubmit`
rpc call in etcdraft chain, due to thread-unsafe of gRPC.

This is going to be tested by subsequent benchmark tests.
Without this serialization, loading etcdraft orderer with
many parallel requests would most likely cause a dead-lock
in communication, which causes benchmark test to stuck.

Change-Id: Id2f6e8f86f7465880725e1aecc7374c3e9e4106e
Signed-off-by: Jay Guo <guojiannan1101@gmail.com>