regenerate ledger testdata using release-1.1
we need to regenerate the testdata using fabric-release v1.1 with the full
definition of collection configs including MembersOrgsPolicy. This is necessary
as one of the stateUpdate listener (i.e., collElgNotifier) compares the existing
MembersOrgsPolicy with the new MembersOrgsPolicy (during upgrade tx) to perform
certain tasks in the ledger. As the existing test data does not contain
MembersOrgsPolicy in the used collectionConfigPkg, it leads to nil pointer
exception.

This CR creates a new testdata with fabric-release v1.1 using
https://gerrit.hyperledger.org/r/#/c/22749/ (which is updated to include
the MembersOrgsPolicy in the collectionConfigPkg)

This CR also enables the v11 ledger test.

FAB-12969 #done

Change-Id: I3d8353145ca7f753f82805b7802d7a68fde9d638
Signed-off-by: senthil <cendhu@gmail.com>
Signed-off-by: manish <manish.sethi@gmail.com>