Prevent verbose error reporting from grpc logging
By default, for errors that implement fmt.Formatter, zap will add a
verboseError field that includes the '%+v' formatted error in the log
record. This is really cool and useful but it can be noisy.

In order to prevent this noise in the grpc interceptors, we can embed the
errors we're logging in a struct; the result is that the struct
satisfies the error interface without implementing fmt.Formatter.

FAB-12953 #done

Change-Id: I312a70a0bc0c0ac88765de319ba67c8ee8c1e3cd
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>