FAB-12983 capability V2_0 for Kafka2RaftMigration
Add orderer capability V2_0 for Kafka2RaftMigration

This capability defines whether the orderer supports a kafka to
raft migration. A kafka-based Ordering Service Node requires
this in order to receive and process a config update with
consensus-type migration commands. Migration is supported from
Kafka to Raft only. If not present, these config updates will
be rejected.

Change-Id: I3b56dec21f0893d0b1df5db30973b4762aab5575
Signed-off-by: Yoav Tock <tock@il.ibm.com>