[FAB-13000] Release resources in token transactor
The token transactor needs to call done on the QueryExecutor, which
implements the token LedgerReader interface, when it is done with it.
In case this is not done, the resources are not released, and timeouts
will occur. This patch adds a Done function to the token LedgerReader
interface, which is called when a token Transactor instance is no
longer needed.

Change-Id: Ie064ea227fb06c3b334acbcd8bc97de210646a26
Signed-off-by: Mathias Bjoerkqvist <mbj@zurich.ibm.com>