[FAB-13054] Prevent int overflow in IsReplicationNeeded
If the OSN is booted without a ledger, the height is 0.
When deciding whether replication is needed, we comapre
the numbers of the last committed block in the system channel.

The height is always the last sequence + 1, therefore
to get the last committed block number of the system channel
we subtract 1 from the height, and when the height is 0-
an integer overflow occurs.

This change set fixes this, and adds a test case.

Change-Id: Ia4718fb206670e50552f61168c73165e27e0fc68
Signed-off-by: yacovm <yacovm@il.ibm.com>