[FAB-13239] terminate container streaming output loop
When vm.docker.attachStdout is enabled, output from chaincode containers
is written to the peer log. A recent change went in that broke the error
handling behavior and resulted in a tight loop of reading from a closed
reader and issuing an error message. This change fixes the error
handling and back-fills test.

Change-Id: Icda853dba90b873f2fbddde990c5f61f7834f1c9
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>