FAB-13265 migration status in channelconfig
Change the channelconfig of an Orderer to reflect the extension of ConsensusType:
- MigrationState
- MigrationContext

Add a method to the bundle to validate the migartion steps of
a new versus old config.

Add test-cases to bundle_test.go to unit-test said method.

Improved comments language.

Needed to regenerate and update mocks in 'common' and 'blockcutter' packages
for unit tests to build correctly.

Change-Id: If060c05bcb9a0e0ca81b1f754a2b0e69a7f6c896
Signed-off-by: Yoav Tock <tock@il.ibm.com>