[FAB-12443] Tolerate CouchDB 2.2 Changes
CouchDB 2.2 does not throw an exception anymore
if a specified index does not exist.
This is a change and unit tests will need to not check
for exceptions if a named index in the query does not exist.

Example:
{"selector":\{"owner":"tom"} ,
"use_index":["indexOwnerDoc","indexOwner"]}

If the document indexOwnerDoc or the index indexOwner does not exist,
an exception will not be thrown.

CouchDB unit tests are changed to reflect the changes.

Change-Id: I8b5259ffb0d55d5e4ae85a49dbf06e58673a0e89
Signed-off-by: David Enyeart <enyeart@us.ibm.com>