FAB-13669 consensus migration: kafka2raft green path #4
This is the fourth of four (4/4) sub-tasks that focus on
the "green" path of consensus-type migration from Kafka to Raft.

By "green" we mean that there are no failures or aborts along the
way. The 4 sub-tasks are staged in a way that minimizes dependencies
between them.

In this sub-task we introduce changes to the etcd/raft-base OSNs such
that they can restart from a ledger that was started as Kafka, migrated,
and restarted. This change concludes all the changes needed to implement
the green path on the "Raft" side.

See respective JIRA item for further details.

Change-Id: I5b408e1cfcb8cf42c39bed4df6c5496792175ef0
Signed-off-by: Yoav Tock <tock@il.ibm.com>