FAB-13666 consensus migration: kafka2raft green path #3
This is the third of four (3/4) sub-tasks that focus on
the "green" path of consensus-type migration from Kafka to Raft.

By "green" we mean that there are no failures or aborts along
the way. The 4 sub-tasks are staged in a way that minimizes
dependencies between them.

In this sub-task we introduce changes to the Kafka-base OSNs
such that they implement the entire green path, until they commit
migration and are ready to be restarted. This
change concludes all the changes needed to implement the green
path on the "Kafka" side.

See respective JIRA item for further details.

Note: some of the mocks had to be re-generated by couterfeiter to make the
unit tests build.

Change-Id: I2747f7d8017c344e9ff3bdd9dd98cbaa1480083f
Signed-off-by: Yoav Tock <tock@il.ibm.com>