FAB-13665 consensus migration: kafka2raft green path #2
This is the second of four (2/4) sub-tasks that focus on
the "green" path of consensus-type migration from Kafka to Raft.

By "green" we mean that there are no failures or aborts along
the way. The 4 sub-tasks are staged in a way that minimizes
dependencies between them.

In this sub-task we introduce a new package:
orderer/consensus/migration. In this package we introduce the
migration Status, which hold the state and context of each chain,
and the migration Stepper, which enforces the migration state machine
in the Kafka chain, post-ordering (i.e. after messages are consumed
from Kafka). It also defines the interface of the migration Controller
which is implemented by the Registrar.

See respective JIRA item for further details.

Change-Id: Ifaad95bcf3bc8fb889ab5f65e817bfe1ebdfa771
Signed-off-by: Yoav Tock <tock@il.ibm.com>