[FAB-13298] Fix TestConfigureClusterListener in MacOS
Some people have port 5000 used in MacOS.
When they try to run unit tests locally,
the test TestConfigureClusterListener fails due to
the port being already in use, which changes the actual
error returned, and as a result - the test fails because
the error string doesn't match what the test expects it to be.

Fixed by selecting a random un-used port.

Change-Id: I687ae38b8c0200e8fb899a519611b50e19cafe08
Signed-off-by: yacovm <yacovm@il.ibm.com>