[FAB-13299] Onboarding: Skip committing existing blocks
If we onboard a new OSN, and crash in the middle-  we start with a ledger
that has some blocks, in some channels.

The onboarding phase, has to fetch all blocks in order to verify the hash
chain and verify the blocks, but we should skip committing blocks that we
already committed before.

This change set implements that, and adjusts the tests accordingly.

Change-Id: I9ffc56cd121e11ddfe675894ae9378fd52157dbe
Signed-off-by: yacovm <yacovm@il.ibm.com>