recon: at a time, only 1 func. can use the Cache
Among ValidateAndPrepare(), PrepareExpiringKeys(), and
RemoveStaleAndCommitPvtDataOfOldBlocks(), we can allow only one
function to execute at a time. The reason is that each function calls
LoadCommittedVersions() which would clear the existing entries in the
transient buffer/cache and load new entries (such a transient buffer/cache
is not applicable for the golevelDB). As a result, these three functions
can interleave and nullify the optimization provided by the bulk read API.
Once the ledger cache (FAB-103) is introduced and existing
LoadCommittedVersions() is refactored to return a map, we can allow
these three functions to execute parallely.

This CR ensures that these functions are not executed parallely.

Further, when we commit the old pvtdata, some of these pvtdata might be
expiring at the next block commit. As the toPurgeList might have been already
constructed by the last Commit() call, we need to update the toPurgeList
with eliglble old pvtdata if it is expiring. Till now, we created new
toPurgeList by calling PrepareExpiringKeys() from
RemoveStaleAndCommitPvtDataOfOldBlocks().

This CR instead updates the existing toPurgeList rather can calling
PrepareExpiringKeys(). This is because, we don't want to execute these
functions in parallel. Also, it is optimal to update the existing
expiring keys list rather than creating a new one especially when no
new regular blocks are being committed for a longer duration and a lot
of old pvtData are being committed.

FAB-13328 #done
FAB-13329 #done

Change-Id: Ia715bd1a376c6ce19caa7af6f6510e0cf6f5a1dd
Signed-off-by: senthil <cendhu@gmail.com>