[FAB-1337] Raft: Commit genesis blocks for non-members
Currently (before this CR), an onboarded OSN doesn't pull any blocks
for channels it doesn't participate in.

Aas a result, when the OSN starts up after onboarding - it doesn't have these
channels in its registrar, and therefore may classify channel
creation transactions differently than its fellow OSNs that do have
the channels it didn't pull.

In order to avoid a state fork, this change set makes the OSN
to commit the genesis block for channels it doesn't participate in.

This is *NOT* done by pulling the genesis blocks, since the OSN may not have
permissions to do that in the first place, but instead - it creates
a genesis block from the system channel block that has the channel creation
transaction.

This change set also changes the integration test for onboarding to
adjust to the changes, namely - ensures the OSN committed the genesis
block for a channel it doesn't participate in, and upon Broadcast,
returns an answer stating it doesn't participate in the channel.

Also, it reduces the run time for the integration test to 50s.

Change-Id: Icf5754df6cedb7725c4d7091c7366ce0b17ff1b7
Signed-off-by: yacovm <yacovm@il.ibm.com>
Signed-off-by: Artem Barger <bartem@il.ibm.com>