[FAB-13360] Fix an etcdraft flaky UT
This CR removes a redundant assertion in etcdraft UT
to avoid flakiness.

The assertion being removed checks that expected number
of MsgApp are dropped since node is disconnected from
network. However, when a raft candidate is elected as
leader, it broadcasts MsgApp containing empty data to
followers, and sometimes this is being counted as part
of dropped MsgApp, which causes unmatched expectation.

We could be more precise by inspecting MsgApp on wire,
and neglect empty messages, so we still perform this
assertion. But it is anyway a redundant check, therefore
can be safely removed.

This CR also further reduces total test time.

Change-Id: I74af9fecfebe20e44c6736a644352f8b67b624e3
Signed-off-by: Jay Guo <guojiannan1101@gmail.com>