[FAB-13427] Make replication tests not depend on time
Sometimes the machine that runs the test, is stuck for too long,
and it makes the client side time out, thus the server side mock
gets out of sync with the tested client side (the replication code).

A way to fix it, is to specify a really big timeout (1 hour),
and in cases where we expect the server not to respond - instead of
making it time out, we make it send an EOF downstream, which will
indicate a failure to the client.

Change-Id: I06ed5bd4a645ae8ace90542fe56d19254bfa42b7
Signed-off-by: yacovm <yacovm@il.ibm.com>