[FAB-13441] Properly capture OSN output
The reconfiguration and onboarding integration tests ensures
that the OSNs stop logging errors at the end of the test,
in order to ensure there aren't any not noticed faults
that occurred due to reconfiguration/onboarding.

The function used the wrong method to obtain a buffer
that is used to read the process's output.

Change-Id: Ieadae1bb083454b195cbfe52b41582dc9dbbf80a
Signed-off-by: yacovm <yacovm@il.ibm.com>